snapcraft.yaml: simplify docker-support plug, and remove network-control
network-control triggers a device cgroup to be enforced for the snap, which breaks docker from being able to start containers, failing with a message like:
```
 [INFO] - Starting module edgeAgent...
[WARN] - Could not start module edgeAgent
[WARN] -         caused by: OCI runtime create failed: container_linux.go:348: starting container process caused "process_linux.go:402: container init caused \"process_linux.go:385: running prestart hook 0 caused \\\"fork/exec /proc/8797/exe: permission denied\\\"\"": unknown
```

Not plugging network-control allows container creation to succeed.

Also, since we are specifying any properties for the docker-support plug, we can just plug it directly from dockerd and not define a separate plug for it.

Finally, disable dockerd on install so that the snap can be installed without the interfaces connected (until we publish in the global store and have auto-connection rules)