snappy-apparmor-tweaks.patch: fix typo in peer group name
Previously, when we ran "ps -o unit= $PID" to get the systemd service name for this process, the output would include a "\n" at the end, and so the trimming of the ".service" at the end wouldn't work. We have to trim the ".service" and the "\n" at the end because AppArmor is strict about the peer group label.