many: replace $SNAP_DATA/run with $SNAP_COMMON/run
The reason for this is that when we initially configure the config scripts for where to put the sockets, $SNAP_DATA is tied to a particular revision so upon refreshing, the socket locations  would need to be updated to the new revision. If we put all the sockets in $SNAP_COMMON it's not an issue as $SNAP_COMMON is shared between the revisions and will always be the same.