snapcraft.yaml: add device-specific interfaces to plugs list
Note that some of these interfaces will trigger the device cgroup to be
turned on when connected, and as such will cause issues when dockerd
creates cgroups for the containers.

The fix for this is to make the docker-support interface upstream
disable turning on a device cgroup when an attribute is defined, much
like the `privileged-containers` attribute currently defined for the
docker-support interface.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>